Bolt on OptArgs (original) API
I've just done this to see what it looks like, but I will shortly
revert this change. I don't think that the OptArgs and OptArgs2 APIs
can be cleanly implemented in the same module, because of the Bool/Flag
change. All of OptArgs Bool options will be shown as --[no-]bool which
is a rather obvious, horrible change if you aren't used to seeing it.

One could possibly introduce a (global?) variable somewhere but that is
the the kind of hack that OptArgs2 is replacing. In any event, it will
be in the history here if it needed at some point.