Rework @CARP_NOT everywhere
The goal has always been to ensure that errors by the caller bubble up
to their script. But I've always fought with Carp about making that work
the way I want it to. This commit is the nuclear option: no errors will
ever be reported as occuring in OptArg2 itself. That could well be
improved by someone else with fresher eyes.