Re-work optargs parsing for hidden SubCmds
This change is in support of sub-commands that exist, but do not have a
parent SubCmd. For example:

    cmd push SRC DEST...           # SRC and DEST are type 'Str'
      cmd push org PATH LOCATION     # But push has subcommands
      cmd push topic ID PATH...

Previously, "push org" and "push topic" would not be detected, because
SRC is not of type "SubCmd". With this change they now are, and the
overall parse logic has also been cleaned up a bit.

I'm not sure that this type of optarg arrangement (with effectively
hidden subcommands) is useful outside my specific use case, but here it
is.