Split out (most) packages into separate files
Originally packing everything into one file was a case of premature
optimization. I felt on my old laptop it might have made a difference,
but today that is certainly not the case. Putting packages into their
own files makes the code easier to understand, and has already exposed
the need for some refactoring.